Add Commander. Users can now supply a PR title
