+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "gh-templating",
-    "version": "1.2.0",
+    "version": "1.3.0",
     "description": "",
     "main": "index.js",
     "bin": {
@@ -32,6 +32,7 @@
     },
     "homepage": "https://github.com/jorgevds/gh-templating#readme",
     "dependencies": {
+        "commander": "^10.0.0",
         "inquirer": "^9.1.4"
     },
     "type": "module"
